[REFACTOR] move review progress from assessors to assessment app
