+++ assessor/views.py
@@ -4,8 +4,7 @@ from rest_framework.response import Response
 from rest_framework.decorators import api_view, permission_classes
 from .services.dashboard import (
     get_all_active_assessees,
-    get_assessor_assessment_events,
-    get_assessee_progress_on_assessment_event
+    get_assessor_assessment_events
 )
 
 
@@ -45,20 +44,3 @@ def serve_get_all_active_assessees(request):
     request_data = request.GET
     active_assessees = get_all_active_assessees(request_data, user=request.user)
     return Response(data=active_assessees)
-
-
-@require_GET
-@api_view(['GET'])
-@permission_classes([IsAuthenticated])
-def serve_get_assessee_progress_on_event(request):
-    """
-    This view will serve as the end point for assessors to get assessment event progress of an assessee
-    ----------------------------------------------------------
-    request-param must contain:
-    assessment-event-id: string
-    assessee-email: string
-    """
-    request_data = request.GET
-    progress_data = get_assessee_progress_on_assessment_event(request_data, user=request.user)
-    return Response(data=progress_data)
-
